requirements.txt
py-cord
python-dotenv
pymongo
bs4
lxml
requests